- bad request should be 400
- Ensure to throw on mismatch default values
- make js parameter api more async
- add a specific confirm delete popup
- create parameter can set default #3848
- enable delete parameters
- set min arg id length to 4 (5 too high)
- protect parameters/products from being deleted
- lint and fix warnings
- add a container for public content
- use a storage acc. for images and favicons

Related work items: #3848

+++ app/web/Controllers/Helpers/CommonEntityControllerBase.cs
@@ -11,9 +11,10 @@ namespace SignalBox.Web.Controllers
     [Authorize]
     [ApiController]
     [Produces("application/json")]
-    public class CommonEntityControllerBase<T> : SignalBoxControllerBase where T : CommonEntity
+    public abstract class CommonEntityControllerBase<T> : SignalBoxControllerBase where T : CommonEntity
     {
         protected ICommonEntityStore<T> store { get; }
+        protected abstract Task<(bool, string)> CanDelete(T entity);
 
         public CommonEntityControllerBase(ICommonEntityStore<T> store)
         {
@@ -82,9 +83,18 @@ namespace SignalBox.Web.Controllers
         [HttpDelete("{id}")]
         public virtual async Task<DeleteResponse> DeleteResource(long id)
         {
-            var result = await store.Remove(id);
-            await store.Context.SaveChanges();
-            return new DeleteResponse(id, Request.Path.Value, result);
+            var entity = await store.Read(id);
+            var (canDelete, message) = await CanDelete(entity);
+            if (canDelete)
+            {
+                var result = await store.Remove(id);
+                await store.Context.SaveChanges();
+                return new DeleteResponse(id, Request.Path.Value, result);
+            }
+            else
+            {
+                throw new BadRequestException("Delete error", message);
+            }
         }
     }
 }
